+++ cognite/powerops/client/_generated/assets/__init__.py
@@ -1,3 +1,0 @@
-from cognite.powerops.client._generated.assets._api_client import PowerAssetAPI
-
-__all__ = ["PowerAssetAPI"]