+++ cognite/powerops/client/powerops_client.py
@@ -5,6 +5,7 @@ import os
 import sys
 from pathlib import Path
 from string import Template
+from typing import Literal
 
 import yaml
 from cognite.client import CogniteClient, global_config
@@ -12,14 +13,9 @@ from cognite.client import CogniteClient, global_config
 from cognite.powerops.client.shop.cogshop_api import CogShopAPI
 from cognite.pygen.utils.external_id_factories import ExternalIdFactory
 
-from ._generated.afrr_bid import AFRRBidAPI
-from ._generated.assets import PowerAssetAPI
-from ._generated.cogshop1 import CogShop1Client
-from ._generated.day_ahead_bid import DayAheadBidAPI
 from ._generated.v1 import PowerOpsModelsV1Client
 from ._generated.v1.data_classes._core import DomainModelWrite
 from .data_set_api import DataSetsAPI
-from .shop.dayahead_trigger_api import DayaheadTriggerAPI
 from .shop.shop_run_api import SHOPRunAPI
 
 # max_domain = max_total (255) - uuid (32) + separator (1)  noqa: ERA001
@@ -35,18 +31,13 @@ class PowerOpsClient:
         read_dataset: str,
         write_dataset: str,
         monitor_dataset: str | None = None,
+        cog_shop_service: Literal["prod", "staging"] | None = None,
     ):
         self.cdf = client
         self.datasets = DataSetsAPI(self.cdf, read_dataset, write_dataset, monitor_dataset)
-        self.cog_shop1 = CogShop1Client(self.cdf)
-        self.assets = PowerAssetAPI(self.cdf)
-        self.afrr_bid = AFRRBidAPI(self.cdf)
-        self.day_ahead_bid = DayAheadBidAPI(self.cdf)
         self.shop = SHOPRunAPI(self.cdf, self.datasets.write_dataset_id)
-        self.workflow = DayaheadTriggerAPI(self.cdf, self.datasets.write_dataset_id)
-
         self.v1 = PowerOpsModelsV1Client(self.cdf)
-        self.cogshop = CogShopAPI(self.cdf, self.v1)
+        self.cogshop = CogShopAPI(self.cdf, self.v1, cog_shop_service)
 
         DomainModelWrite.external_id_factory = ExternalIdFactory.create_external_id_factory(
             prefix_ext_id_factory=ExternalIdFactory(